moving new_signals into signals
